wallet incryption refactor
